re-order collection sync spec
